Add props challenge solution
The rendered HTML coming out is omitting
a semi-colon for the inline css styles , so I
had to include one in the object prop string.

This is considered bad style and I am now getting a warning but
it was the only way I could see to pass the learnyoureact 'verify'.